// monorepo/apps/web/src/modules/warp/packages/client/features/form/types.ts
export type FormFieldWithChildrenType = FormField & {
  children?: FormFieldWithChildrenType[];
};

export type FormFieldOptionsType =
  | "displayOptions"
  | "interfaceOptions"
  | "fieldOptions";

export type InfoIconPropsType = {
  content: "";
  popoverProps?: PopoverProps;
  popoverDropdownProps?: PopoverDropdownProps;
};

export type FormFieldInterfaces = {
// ...
  monthyear: {
    fieldOptions: {
      label: string;
      required: boolean;
      enable: boolean;
      readonly: boolean;
    };
    interfaceOptions: {
      showLabel: boolean;
      title: string;
      placeholder?: string;
      isHeading?: boolean;
      headingSize: string;
      showTitleDivider?: boolean;
      infoIconProps?: InfoIconPropsType;
      subtitle: string;
      maxLength?: number;
      maxValue?: number;
      isToolTip: boolean;
      isemail: boolean;
      isAlphabetSpecialChar: boolean;
      isAutoFilled?: boolean;
      isHyperLink?: boolean;
      isAutoCalculate?: boolean;
      isYearOnly?: boolean;
    };
    displayOptions: any;
    displayRules: RuleType<
      Pick<FormFieldInterfaces["monthyear"], FormFieldOptionsType>
    >[];
    validationRules: ValidationRuleType[];
    autoCalculatedCalculation?: ValidationRuleType[];
  };
// ...
};

export type FormFieldInterfaceKeysType = keyof FormFieldInterfaces;

export type FormFieldInterfaceType<T extends FormFieldInterfaceKeysType> = Omit<
  FormField,
  keyof FormFieldInterfaces[T]
> &
  FormFieldInterfaces[T];

export type RuleType<TOptions = any> = {
  name: string;
  rule: string;
} & TOptions;

export type ValidationRuleType<T = any> = RuleType<
  T & {
    message: string;
  }
>;
// ...
export type SimpleFormFieldControlPropsType =
  | {
    isSimple?: boolean;
    name?: string;
    value?: any;
    onChange?: (value: any, isSparkIconClick?: boolean) => void;
    errorMessage?: any;
    rowIndex?: number | undefined;
    ansId?: string | undefined;
    tdheading?: any;
    isStateValuefromAI?: boolean;
    isShowSparkIcon?: boolean;
  }
  | {
    isSimple: boolean;
    value: any;
    name: string;
    onChange: (value: any, isSparkIconClick?: boolean) => void;
    errorMessage?: any;
    rowIndex: number | undefined;
    ansId: string | undefined;
    tdheading: any;
    isStateValuefromAI?: boolean;
    isShowSparkIcon?: boolean;
  };

export type FormFieldControl<
  TFormField extends FormFieldInterfaceKeysType = "default",
  T = any,
> = FC<
  PropsWithChildren<
    {
      formField: Omit<
        FormFieldWithChildrenType,
        keyof FormFieldInterfaces[TFormField]
      > &
      FormFieldInterfaces[TFormField];
    } & SimpleFormFieldControlPropsType &
    T
  >
>;